pyproject.toml
[project]
name = "telebot"
version = "0.1.0"
description = ""
authors = [
    {name = "igoruxa007",email = "igoruxa007@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot (>=22.1,<23.0)",
    "black (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
